LPS-129242 Split translate into small components
