+++ modules/apps/translation/translation-web/src/main/resources/META-INF/resources/js/translate/Translate.js
@@ -14,162 +14,14 @@
 
 import ClayAlert from '@clayui/alert';
 import ClayButton from '@clayui/button';
-import ClayForm, {ClayInput} from '@clayui/form';
-import ClayIcon from '@clayui/icon';
 import ClayLayout from '@clayui/layout';
 import ClayLink from '@clayui/link';
-import {ClassicEditor} from 'frontend-editor-ckeditor-web';
 import PropTypes from 'prop-types';
 import React, {useState} from 'react';
 
 import TranslateLanguagesSelector from './components/TranslateLanguagesSelector';
-
-const noop = () => {};
-
-const TranslationFieldEditor = ({
-	editorConfiguration,
-	id,
-	label,
-	sourceContent,
-	sourceContentDir,
-	targetContent,
-	targetContentDir,
-	onChange = noop,
-}) => {
-	const [value, setValue] = useState(targetContent);
-
-	return (
-		<ClayLayout.Row>
-			<ClayLayout.Col md={6}>
-				<ClayForm.Group>
-					<label className="control-label">{label}</label>
-					<div
-						className="translation-editor-preview"
-						dangerouslySetInnerHTML={{__html: sourceContent}}
-						dir={sourceContentDir}
-					/>
-				</ClayForm.Group>
-			</ClayLayout.Col>
-			<ClayLayout.Col md={6}>
-				<ClayForm.Group>
-					<label className="control-label">{label}</label>
-					<ClassicEditor
-						contents={targetContent}
-						editorConfig={{
-							...editorConfiguration.editorConfig,
-							contentsLangDirection: targetContentDir,
-						}}
-						name={id}
-						onChange={(data) => {
-							if (value !== data.trim()) {
-								setValue(data);
-								onChange(data);
-							}
-						}}
-					/>
-					<input defaultValue={value} name={id} type="hidden" />
-				</ClayForm.Group>
-			</ClayLayout.Col>
-		</ClayLayout.Row>
-	);
-};
-
-const TranslationFieldInput = ({
-	id,
-	label,
-	multiline,
-	sourceContent,
-	sourceContentDir,
-	targetContent,
-	targetContentDir,
-	onChange = noop,
-}) => {
-	const [value, setValue] = useState(targetContent);
-
-	return (
-		<ClayLayout.Row>
-			<ClayLayout.Col md={6}>
-				<ClayForm.Group>
-					<label className="control-label">{label}</label>
-					<ClayInput
-						component={multiline ? 'textarea' : undefined}
-						defaultValue={sourceContent}
-						dir={sourceContentDir}
-						readOnly
-						type="text"
-					/>
-				</ClayForm.Group>
-			</ClayLayout.Col>
-			<ClayLayout.Col md={6}>
-				<ClayForm.Group>
-					<label className="control-label" htmlFor={id}>
-						{label}
-					</label>
-					<ClayInput
-						component={multiline ? 'textarea' : undefined}
-						dir={targetContentDir}
-						id={id}
-						name={id}
-						onChange={(event) => {
-							const data = event.target.value;
-
-							setValue(data);
-							onChange(data);
-						}}
-						type="text"
-						value={value}
-					/>
-				</ClayForm.Group>
-			</ClayLayout.Col>
-		</ClayLayout.Row>
-	);
-};
-
-const TranslationFieldSetEntries = ({
-	infoFieldSetEntries,
-	onChange,
-	portletNamespace,
-}) =>
-	infoFieldSetEntries.map(({fields, legend}) => (
-		<React.Fragment key={legend}>
-			<ClayLayout.Row>
-				<ClayLayout.Col md={6}>
-					<div className="fieldset-title">{legend}</div>
-				</ClayLayout.Col>
-				<ClayLayout.Col md={6}>
-					<div className="fieldset-title">{legend}</div>
-				</ClayLayout.Col>
-			</ClayLayout.Row>
-			{fields.map((field) => {
-				const fieldProps = {
-					...field,
-					id: `${portletNamespace}${field.id}`,
-					onChange,
-				};
-
-				return field.html ? (
-					<TranslationFieldEditor key={field.id} {...fieldProps} />
-				) : (
-					<TranslationFieldInput key={field.id} {...fieldProps} />
-				);
-			})}
-		</React.Fragment>
-	));
-
-const TranslationHeader = ({sourceLanguageIdTitle, targetLanguageIdTitle}) => (
-	<ClayLayout.Row>
-		<ClayLayout.Col md={6}>
-			<ClayIcon symbol={sourceLanguageIdTitle.toLowerCase()} />
-			<span className="ml-1">{sourceLanguageIdTitle}</span>
-			<div className="separator" />
-		</ClayLayout.Col>
-		<ClayLayout.Col md={6}>
-			<ClayIcon symbol={targetLanguageIdTitle.toLowerCase()} />
-			<span className="ml-1">{targetLanguageIdTitle}</span>
-			<div className="separator" />
-		</ClayLayout.Col>
-	</ClayLayout.Row>
-);
+import TranslateFieldSetEntries from './components/TranslateFieldSetEntries';
+import TranslateHeader from './components/TranslateHeader';
 
 const Translate = ({
 	aditionalFields,
@@ -270,11 +122,11 @@ const Translate = ({
 						</ClayAlert>
 					) : (
 						<>
-							<TranslationHeader
+							<TranslateHeader
 								sourceLanguageIdTitle={sourceLanguageIdTitle}
 								targetLanguageIdTitle={targetLanguageIdTitle}
 							/>
-							<TranslationFieldSetEntries
+							<TranslateFieldSetEntries
 								infoFieldSetEntries={infoFieldSetEntries}
 								onChange={() => setFormHaschanges(() => true)}
 								portletNamespace={portletNamespace}
